demo/requirements.txt
# LocalMind Collective - Oxford AI Summit 2025 Demo Requirements

# Core Framework
fastapi
uvicorn[standard]
websockets

# HTTP and Async
aiohttp
httpx

# Configuration and Validation
pydantic
pydantic-settings
python-dotenv

# Logging and Monitoring
loguru
psutil

# Optional: Enhanced monitoring (uncomment if needed)
# structlog>=23.2.0
# PyYAML>=6.0.1
# prometheus-client>=0.19.0
# opentelemetry-api>=1.21.0

#PII Demo
anthropic
pandas
numpy
mcp
aiohttp